+++ pkg/middleware/corsMW.go
@@ -1,0 +1,18 @@
+package middleware
+
+import (
+	"mail/config"
+	"net/http"
+	"slices"
+)
+
+func CORS(next http.Handler, cfg *config.Config) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		origin := r.Header.Get("Origin")
+		if slices.Contains(cfg.HTTPServer.AllowedIPsByCORS, origin) {
+			w.Header().Set("Access-Control-Allow-Origin", origin)
+			w.Header().Add("Vary", "Origin")
+		}
+		next.ServeHTTP(w, r)
+	})
+}
